requirements.txt
psycopg2-binary
seleniumbase
sqlalchemy
cloudinary
requests
python-dotenv
pandas